+++ packages/fury-adapter-oas3-parser/lib/parser/oas/parseOperationObject.js
@@ -11,12 +11,12 @@ const pipeParseResult = require('../../pipeParseResult');
 const parseObject = require('../parseObject');
 const parseString = require('../parseString');
 const parseResponsesObject = require('./parseResponsesObject');
+const parseParameterObjects = require('./parseParameterObjects');
 
 const name = 'Operation Object';
 const requiredKeys = ['responses'];
 const unsupportedKeys = [
-  'tags', 'externalDocs', 'parameters', 'requestBody',
-  'callbacks', 'deprecated', 'security',
+  'tags', 'externalDocs', 'requestBody', 'callbacks', 'deprecated', 'security',
 ];
 const isUnsupportedKey = R.anyPass(R.map(hasKey, unsupportedKeys));
 
@@ -33,16 +33,40 @@ function createTransactions(namespace, member, operation) {
   });
 }
 
+function hrefVariablesFromParameters(namespace, parameters) {
+  const path = parameters.get('path')
+    ? parameters.get('path')
+    : new namespace.elements.HrefVariables();
+
+  const query = parameters.get('query')
+    ? parameters.get('query')
+    : new namespace.elements.HrefVariables();
+
+  if (!path.isEmpty || !query.isEmpty) {
+    return path.concat(query);
+  }
+
+  return undefined;
+}
+
+function hrefFromParameters(path, queryParameters) {
+  const href = path.clone();
+  const queryString = queryParameters.keys().join(',');
+  href.content += `{?${queryString}}`;
+  return href;
+}
+
 /**
  * Parse Operation Object
  *
  * @param namespace {Namespace}
+ * @param path {StringElement}
  * @param element {Element}
  * @returns ParseResult<Transition>
  *
  * @see https://github.com/OAI/OpenAPI-Specification/blob/master/versions/3.0.0.md#operationObject
  */
-function parseOperationObject(context, member) {
+function parseOperationObject(context, path, member) {
   const { namespace } = context;
 
   const isUnique = element => context.registerId(element.toValue());
@@ -60,6 +84,7 @@ function parseOperationObject(context, member) {
     [hasKey('description'), parseCopy(context, name, false)],
     [hasKey('operationId'), pipeParseResult(namespace, parseString(context, name, false), parseOperationId)],
     [hasKey('responses'), R.compose(parseResponsesObject(context), getValue)],
+    [hasKey('parameters'), R.compose(parseParameterObjects(context, name), getValue)],
 
     [isUnsupportedKey, createUnsupportedMemberWarning(namespace, name)],
 
@@ -85,6 +110,16 @@ function parseOperationObject(context, member) {
       const description = operation.get('description');
       if (description) {
         transition.push(description);
+      }
+
+      const parameters = operation.get('parameters');
+      if (parameters) {
+        const queryParameters = parameters.get('query');
+        if (queryParameters) {
+          transition.href = hrefFromParameters(path, queryParameters);
+        }
+
+        transition.hrefVariables = hrefVariablesFromParameters(namespace, parameters);
       }
 
       const transactions = createTransactions(namespace, member, operation);